collect_data/requirements.txt
python-dotenv==0.14.0
bleak==0.9.1